generator initial_state merges the individual processes initial_state
